Fix the Data Being Sent to WebAssembly so Different Numbers are Returned
Shrink canvas down so the data is the right length just to see if it works
Since the canvas only has black, RGB are all 0 and only Alpha has value so we have to copy that value.
The returned values from WebAssembly look mostly nonsense, but the brush size is also really large right now.